refactor: index 누락 추가, import 경로 변경
