+++ src/components/templates/EventGridTemplate/index.ts
@@ -1,0 +1,3 @@
+import EventGridTemplate from './EventGridTemplate';
+
+export { EventGridTemplate };
